UI refactor  and added time series

+++ src/main/java/pl/edu/pwr/chrono/readmodel/dto/SentenceWordCount.java
@@ -2,17 +2,14 @@ package pl.edu.pwr.chrono.readmodel.dto;
 
 import lombok.Data;
 
-/**
- * Created by tnaskret on 10.02.16.
- */
 @Data
 public class SentenceWordCount {
 
     private Integer sentenceId;
-    private Integer wordCount;
-    private Integer letterCount;
+    private Long wordCount;
+    private Long letterCount;
 
-    public SentenceWordCount(Integer sentenceId, Integer wordCount, Integer letterCount) {
+    public SentenceWordCount(Integer sentenceId, Long wordCount, Long letterCount) {
         this.sentenceId = sentenceId;
         this.wordCount = wordCount;
         this.letterCount = letterCount;
